# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.10",
    "title": "META API",
    "description": "API文件"
  },
  "host": "six-northnode.onrender.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "jwt token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/admin/signup": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "管理員註冊",
        "description": "管理員註冊",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "會員註冊用",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "name": {
                  "type": "string",
                  "example": "name"
                }
              },
              "required": [
                "email",
                "password",
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "name"
                    },
                    "email": {
                      "type": "string",
                      "example": "example@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$12$ZNPx7mwkP0/QC6h6eaW5eeuJXdEQrolyIejEZEc55p2doi3C6g/3q"
                    },
                    "role": {
                      "type": "string",
                      "example": "admin"
                    },
                    "_id": {
                      "type": "string",
                      "example": "644f92cfcb3582d71cee8d80"
                    },
                    "createdAt": {
                      "type": "string",
                      "example": "2023-05-01T10:22:07.078Z"
                    }
                  }
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0NGY5MmNmY2IzNTgyZDcxY2VlOGQ4MCIsImlhdCI6MTY4MjkzNjUyNywiZXhwIjoxNjgzNTQxMzI3fQ.yt4zPWEnxZ5Vj49Mcsq61j1qo4Bv0pfdH8GbJazkyE0"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "管理員登入",
        "description": "管理員登入",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": "true",
            "description": "會員註冊用",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回傳範例資料",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "string",
                  "example": "登入成功"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}